+++ test/test_hypothesis/test_connection.py
@@ -1,15 +1,19 @@
-from .base import (
-    BaseTest,
-    commands,
-    values,
-    st,
-    dbnums,
-    common_commands,
+import pytest
+
+from .base import BaseMachine, commands, dbnums, run_machine, st, values
+
+connection_commands = (
+    commands(st.just("echo"), values)
+    | commands(st.just("ping"), st.lists(values, max_size=2))
+    | commands(st.just("select"), dbnums)
 )
 
 
-class TestConnection(BaseTest):
-    # TODO: tests for select
-    connection_commands = commands(st.just("echo"), values) | commands(st.just("ping"), st.lists(values, max_size=2))
-    command_strategy_redis_only = commands(st.just("swapdb"), dbnums, dbnums)
-    command_strategy = connection_commands | common_commands
+class ConnectionMachine(BaseMachine):
+    base_commands = connection_commands
+    redis_only_commands = commands(st.just("swapdb"), dbnums, dbnums)
+
+
+@pytest.mark.slow
+def test_connection(hypothesis_config):
+    run_machine(ConnectionMachine, hypothesis_config)